+++ device/general/unenroll-updatable-assets.ps1
@@ -7,20 +7,19 @@
 
   .NOTES
   Permissions (Graph):
-  - Device.Read.All
   - WindowsUpdates.ReadWrite.All
 
-  .PARAMETER DeviceName
-  Device Name of the device to unenroll.
+  .PARAMETER DeviceId
+  DeviceId of the device to unenroll.
 
   .PARAMETER UpdateCategory
-  Category of updates to unenroll from. Possible values are: driver, feature, quality.
+  Category of updates to unenroll from. Possible values are: driver, feature, quality or all (delete).
 
   .PARAMETER CallerName
   Caller name for auditing purposes.
 
   .INPUTS
-  DeviceName, UpdateCategory, and CallerName
+  DeviceId, UpdateCategory, and CallerName
 #>
 
 #Requires -Modules @{ModuleName = "RealmJoin.RunbookHelper"; ModuleVersion = "0.8.3" }
@@ -31,8 +30,8 @@ param(
     [Parameter(Mandatory = $true)]
     [string] $DeviceId,
     [Parameter(Mandatory = $true)]
-    [ValidateSet("driver", "feature", "quality")]
-    [string] $UpdateCategory
+    [ValidateSet("driver", "feature", "quality", "all")]
+    [string] $UpdateCategory = "all"
 )
 
 Write-RjRbLog -Message "Caller: '$CallerName'" -Verbose
@@ -42,11 +41,10 @@ Connect-RjRbGraph -Force
 function Unenroll-Device {
     param (
         [string]$DeviceId,
-        [string]$DeviceName,
         [string]$UpdateCategory
     )
 
-    Write-RjRbLog -Message "Unenrolling device: $DeviceName (ID: $DeviceId) from $UpdateCategory updates" -Verbose
+    Write-Output "Unenrolling device with ID $DeviceId from $UpdateCategory updates"
 
     $unenrollBody = @{
         updateCategory = $UpdateCategory
@@ -59,32 +57,23 @@ function Unenroll-Device {
     } 
 
     try {
-        $unenrollResponse = Invoke-RjRbRestMethodGraph -Resource "/admin/windows/updates/updatableAssets/unenrollAssets" -Method POST -Body $unenrollBody -Beta
-        Write-Output "Device '$DeviceName' (ID: $DeviceId) successfully unenrolled from $UpdateCategory updates."
+        $unenrollResponse = $null
+        if($UpdateCategory -eq "all") {
+            $unenrollResponse = Invoke-RjRbRestMethodGraph -Resource "/admin/windows/updates/updatableAssets/$DeviceId" -Method DELETE -Beta
+            Write-Output "- Triggered unenroll from updatableAssets via deletion."    
+        } else {
+            $unenrollResponse = Invoke-RjRbRestMethodGraph -Resource "/admin/windows/updates/updatableAssets/unenrollAssets" -Method POST -Body $unenrollBody -Beta
+            Write-Output "- Triggered unenroll from updatableAssets for category $UpdateCategory."
+        }
+        if(!$unenrollResponse) {
+            Write-Output "- Note: Empty Graph response (device probably already offboarded)"
+        }
     }
     catch {
-        $errorResponse = $_.ErrorDetails.Message | ConvertFrom-Json
-        Write-Output "Failed to unenroll device '$DeviceName' (ID: $DeviceId)."
-        Write-Output "Error: $($errorResponse.error.message)"
+        $errorResponse = $_
+        Write-Output "- Failed to unenroll device."
+        Write-Output "- Error: $($errorResponse)"
     }
 }
 
-Write-RjRbLog -Message "Unenrolling device: $DeviceId" -Verbose
-"## Unenrolling device: $DeviceId"
-
-# Get Device ID from Microsoft Entra based on Device Name
-Write-RjRbLog -Message "Fetching Device Details for $DeviceId" -Verbose
-
-$deviceResponse = Invoke-RjRbRestMethodGraph -Resource "/devices" -OdFilter "displayName eq '$DeviceId'" -ErrorAction SilentlyContinue
-
-if ($deviceResponse) {
-    $deviceName = $deviceResponse.displayName
-    Write-RjRbLog -Message "Device $deviceName Found!" -Verbose
-    Unenroll-Device -DeviceId $DeviceId -DeviceName $DeviceName -UpdateCategory $UpdateCategory
-}
-else {
-    "Device not found: $DeviceId"
-    throw "DeviceId $DeviceId not found."
-}
-
-
+Unenroll-Device -DeviceId $DeviceId -UpdateCategory $UpdateCategory